requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
blinker==1.9.0
boto3==1.40.25
botocore==1.40.25
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
Flask==3.1.2
Flask-Dance==7.1.0
Flask-Login==0.6.3  # Required by Flask-Security-Too
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
MarkupSafe==3.0.2
openai==1.106.1
packaging==25.0
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
s3transfer==0.13.1
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
URLObject==3.0.0
Werkzeug==3.1.3
WTForms==3.2.1
